frontend + fix add products as admin

+++ ComputerStore/src/main/java/com/example/ComputerStore/model/Processor.java
@@ -7,12 +7,12 @@ import jakarta.persistence.Table;
 @Table(name = "processors")
 public class Processor extends Product {
     private Integer coreCount;
-    private Integer coreClock;
+    private Float coreClock;
     private String socket;
 
     public Processor() {}
 
-    public Processor(Integer productId, String name, float price, Integer coreCount, Integer coreClock, String socket) {
+    public Processor(Integer productId, String name, float price, Integer coreCount, Float coreClock, String socket) {
         super(productId, name, price);
         this.coreCount = coreCount;
         this.coreClock = coreClock;
@@ -21,8 +21,8 @@ public class Processor extends Product {
 
     public Integer getCoreCount() { return coreCount; }
     public void setCoreCount(Integer coreCount) { this.coreCount = coreCount; }
-    public Integer getCoreClock() { return coreClock; }
-    public void setCoreClock(Integer coreClock) { this.coreClock = coreClock; }
+    public Float getCoreClock() { return coreClock; }
+    public void setCoreClock(Float coreClock) { this.coreClock = coreClock; }
     public String getSocket() { return socket; }
     public void setSocket(String socket) { this.socket = socket; }
 }
